@@ -1,10 +1,10 @@
 name: video_compress
 description: Light library of video manipulation of Flutter. Compress video, remove audio, get video thumbnail from dart code.
-version: 3.1.2
+version: 3.1.3
 homepage: https://github.com/jonataslaw/VideoCompress
 
 environment:
-  sdk: '>=2.16.2 <3.0.0'
+  sdk: ">=3.0.0 <4.0.0"
   flutter: ">=2.0.0"
 
 dependencies:
@@ -14,7 +14,7 @@ dependencies:
 # dev_dependencies:
 #   flutter_test:
 #     sdk: flutter
-  
+
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
 
