@@ -1,8 +1,9 @@
+import 'dart:io';
+
+import 'package:file_selector/file_selector.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:video_compress/video_compress.dart';
-import 'package:file_selector/file_selector.dart';
-import 'dart:io';
 import 'package:video_compress_example/video_thumbnail.dart';
 
 void main() {
@@ -75,7 +76,7 @@ class _MyHomePageState extends State<MyHomePage> {
             ),
             Text(
               '$_counter',
-              style: Theme.of(context).textTheme.headline4,
+              style: Theme.of(context).textTheme.headlineMedium,
             ),
             InkWell(
                 child: Icon(
